feat(tools): move 11ty plugins to pfe-tools
