--- docs/_data/corePackages.cjs
@@ -1,4 +1,0 @@
-module.exports = async function corePackages() {
-  const { getPackageData } = await import('@patternfly/pfe-tools/11ty');
-  return getPackageData('core');
-};
