--- docs/_data/meta.cjs
@@ -1,3 +1,0 @@
-module.exports = {
-  env: process.env.ELEVENTY_ENV
-};
