+++ docs/_includes/layout-demo.njk
@@ -4,6 +4,7 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <meta name="description" content="PatternFly Elements: A set of community-created web components based on PatternFly design.">
+    <meta name="pfe-auto-reveal" content="true">
     <link href="/brand/logo/svg/pfe-icon-blue.svg" rel="shortcut icon">
     <title>{{ element.title }} | PatternFly Elements</title>
     <link rel="preconnect" href="https://ga.jspm.io">
